add tp1_ej1 solution and boolean type

+++ tp1/ej1/tp1_ej1.h
@@ -2,6 +2,6 @@
 #ifndef TP1_EJ1_H_INCLUDED
 #define TP1_EJ1_H_INCLUDED
 
-int tp1_ej1();
+long tp1_ej1();
 
 #endif // TP1_EJ1_H_INCLUDED
